Admin working
- Just done simple functions add update delete.

+++ server-nodejs/model/userPetsModel.js
@@ -38,13 +38,11 @@ const deleteUserPet = async (userPetId, userId, petId) => {
 
 const deleteUserPetbyUserid = async (userId) => {
     try {
-        const result = await db.query(
-            'DELETE FROM user_pets WHERE user_id = $1 RETURNING *',
-            [userId]
-        );
-        return result.rows[0]; // Return the deleted pet's ID if it exists
+        const result = await pool.query('DELETE FROM user_pets WHERE user_id = $1 RETURNING *', [userId]);
+        return result.rows[0];
     } catch (error) {
-        console.error(error);
+        console.error('Error deleting user pets by user ID:', error);
+        throw error;
     }
 };
